+++ test/integration-serial/serial_task_run_test.rb
@@ -3,7 +3,7 @@ require 'integration_test_helper'
 
 class SerialTaskRunTest < Krane::IntegrationTest
   include TaskRunnerTestHelper
-  include StatsDHelper
+  include StatsD::Instrument::Assertions
 
   # Mocha is not thread-safe: https://github.com/freerange/mocha#thread-safety
   def test_run_without_verify_result_fails_if_pod_was_not_created
@@ -35,7 +35,7 @@ class SerialTaskRunTest < Krane::IntegrationTest
     bad_ns = "missing"
     task_runner = build_task_runner(ns: bad_ns)
 
-    metrics = capture_statsd_calls do
+    metrics = capture_statsd_calls(client: Krane::StatsD.client) do
       result = task_runner.run(run_params)
       assert_task_run_failure(result)
     end
@@ -52,7 +52,7 @@ class SerialTaskRunTest < Krane::IntegrationTest
     deploy_task_template
     task_runner = build_task_runner
 
-    metrics = capture_statsd_calls do
+    metrics = capture_statsd_calls(client: Krane::StatsD.client) do
       result = task_runner.run(run_params.merge(verify_result: false))
       assert_task_run_success(result)
     end
@@ -69,7 +69,7 @@ class SerialTaskRunTest < Krane::IntegrationTest
     deploy_task_template
     task_runner = build_task_runner(max_watch_seconds: 0)
 
-    metrics = capture_statsd_calls do
+    metrics = capture_statsd_calls(client: Krane::StatsD.client) do
       result = task_runner.run(run_params.merge(args: ["sleep 5"]))
       assert_task_run_failure(result, :timed_out)
     end