+++ test/unit/krane/statsd_test.rb
@@ -3,7 +3,8 @@
 require 'test_helper'
 
 class StatsDTest < Krane::TestCase
-  include StatsDHelper
+  include StatsD::Instrument::Assertions
+
   class TestMeasureClass
     extend(Krane::StatsD::MeasureMethods)
 
@@ -31,39 +32,6 @@ class StatsDTest < Krane::TestCase
     measure_method :thing_to_measure
   end
 
-  def test_build_when_statsd_addr_env_present_but_statsd_implementation_is_not
-    original_addr = ENV['STATSD_ADDR']
-    ENV['STATSD_ADDR'] = '127.0.0.1'
-    original_impl = ENV['STATSD_IMPLEMENTATION']
-    ENV['STATSD_IMPLEMENTATION'] = nil
-    original_dev = ENV['STATSD_DEV']
-    ENV['STATSD_DEV'] = nil
-
-    Krane::StatsD.build
-
-    assert_equal(:datadog, Krane::StatsD.backend.implementation)
-  ensure
-    ENV['STATSD_ADDR'] = original_addr
-    ENV['STATSD_IMPLEMENTATION'] = original_impl
-    ENV['STATSD_DEV'] = original_dev
-    Krane::StatsD.build
-  end
-
-  def test_kubernetes_statsd_does_not_override_global_config
-    old_prefix = ::StatsD.prefix
-    old_sample_rate = ::StatsD.default_sample_rate
-
-    ::StatsD.prefix = "test"
-    ::StatsD.default_sample_rate = 2.0
-    Krane::StatsD.build
-    refute_equal(Krane::StatsD.prefix, ::StatsD.prefix)
-    refute_equal(Krane::StatsD.default_sample_rate, ::StatsD.default_sample_rate)
-    refute_equal(Krane::StatsD.backend, ::StatsD.backend)
-  ensure
-    ::StatsD.prefix = old_prefix
-    ::StatsD.default_sample_rate = old_sample_rate
-  end
-
   def test_measuring_non_existent_method_raises
     assert_raises_message(NotImplementedError, "Cannot instrument undefined method bogus_method") do
       TestMeasureClass.measure_method(:bogus_method)
@@ -75,7 +43,7 @@ class StatsDTest < Krane::TestCase
   end
 
   def test_measure_method_uses_expected_name_and_tags
-    metrics = capture_statsd_calls do
+    metrics = capture_statsd_calls(client: Krane::StatsD.client) do
       TestMeasureClass.new.thing_to_measure
     end
     assert_predicate(metrics, :one?, "Expected 1 metric, got #{metrics.length}")
@@ -84,7 +52,7 @@ class StatsDTest < Krane::TestCase
   end
 
   def test_measure_method_with_custom_metric_name
-    metrics = capture_statsd_calls do
+    metrics = capture_statsd_calls(client: Krane::StatsD.client) do
       TestMeasureClass.new.measure_with_custom_metric
     end
     assert_predicate(metrics, :one?, "Expected 1 metric, got #{metrics.length}")
@@ -93,16 +61,16 @@ class StatsDTest < Krane::TestCase
   end
 
   def test_measure_method_with_statsd_tags_undefined
-    metrics = capture_statsd_calls do
+    metrics = capture_statsd_calls(client: Krane::StatsD.client) do
       TestMeasureNoTags.new.thing_to_measure
     end
     assert_predicate(metrics, :one?, "Expected 1 metric, got #{metrics.length}")
     assert_equal("KubernetesDeploy.thing_to_measure.duration", metrics.first.name)
-    assert_empty(metrics.first.tags)
+    assert_nil(metrics.first.tags)
   end
 
   def test_measure_method_that_raises_with_hash_tags
-    metrics = capture_statsd_calls do
+    metrics = capture_statsd_calls(client: Krane::StatsD.client) do
       tester = TestMeasureClass.new
       tester.expects(:statsd_tags).returns(test: true)
       assert_raises(ArgumentError) do
@@ -115,7 +83,7 @@ class StatsDTest < Krane::TestCase
   end
 
   def test_measure_method_that_raises_with_array_tags
-    metrics = capture_statsd_calls do
+    metrics = capture_statsd_calls(client: Krane::StatsD.client) do
       tester = TestMeasureClass.new
       tester.expects(:statsd_tags).returns(["test:true"])
       assert_raises(ArgumentError) do