+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 *Important!*
 - The next release will be 1.0.0, which means that master will contain breaking changes.
 
+*Other*
+- Refactor StatsD usage so we can depend on the latest version again.
+
 ## 0.30.0
 
 *Enhancements*